+++ src/env.d.ts
@@ -1,0 +1,3 @@
+/// <reference path="../.astro/types.d.ts" />
+
+declare const __PACKAGE_VERSION__: string;